Backward Facing Step
As we know, the major difficulty encountered during solution of incompressible flow is the nonavailability of the equation of state, which gives us a direct relationship between pressure and temperature. We need to find formulations that can couple pressure with the Navier Stokes equation. Pressure Poisson is one of the methods to solve the problem. There are many algorithms developed to couple Pressure Poisson equation with the Navier Stokes equations. One of them is SIMPLE (Semi Implicit Method for Pressure Correction) originally proposed by Patankar and Spalding [1] in 1972. A revised version of the SIMPLE proposed by Patankar, 1981 which had no gradient of Pressure in the momentum equation. Similarly, SIMPLEC by Doormal and Kaithby are some of the methods developed to couple Pressure Poisson to Navier Stokes equation to solve for incompressible flow. 
The work presented here uses Pressure Poisson equation coupled to Navier Stokes equation to solve laminar flow over a backward facing step of height h. SIMPLE method proposed by Patankar and Spalding 1972 used for pressure and velocity correlations. In addition, a scalar, fourth difference, third order accurate artificial dissipation for stability introduced. To accelerate the convergence rate of the method, local time stepping and the method of residual smoothing used. In the end solutions obtained compared with benchmarked solutions of Armaly et al[2].